+++ .gitignore
@@ -2,3 +2,5 @@
 .vscode/
 target/
 dependency-reduced-pom.xml
+
+.mise/
