+++ roadrunner-core/src/main/java/io/roadrunner/core/Bootstrap.java
@@ -16,17 +16,19 @@
 package io.roadrunner.core;
 
 import io.roadrunner.api.Roadrunner;
-import io.roadrunner.api.latency.LatencyRecorder;
 import io.roadrunner.api.measurments.MeasurementProgress;
 import io.roadrunner.core.internal.ClosedWorldStrategy;
 import io.roadrunner.core.internal.DefaultRoadrunner;
 import io.roadrunner.core.internal.ExecutionStrategy;
 import io.roadrunner.core.internal.OpenWorldStrategy;
+import io.roadrunner.latency.recording.PauseDetectorKind;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Duration;
+import java.util.EnumSet;
+import java.util.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,7 +39,7 @@ public class Bootstrap {
     private ExecutionStrategy strategy;
     private MeasurementProgress measurementProgress = MeasurementProgress.NO_OP;
     private Path outputDir;
-    private LatencyRecorder latencyRecorder = LatencyRecorder.noop();
+    private EnumSet<PauseDetectorKind> pauseDetectors = EnumSet.noneOf(PauseDetectorKind.class);
 
     /**
      * Configure the closed-world load model: N concurrent users each loop until the total
@@ -62,13 +64,13 @@ public class Bootstrap {
         return this;
     }
 
-    public Bootstrap withLatencyRecorder(LatencyRecorder latencyRecorder) {
-        this.latencyRecorder = latencyRecorder;
+    public Bootstrap withOutputDir(Path outputDir) {
+        this.outputDir = outputDir;
         return this;
     }
 
-    public Bootstrap withOutputDir(Path outputDir) {
-        this.outputDir = outputDir;
+    public Bootstrap withPauseDetectorKinds(EnumSet<PauseDetectorKind> pauseDetectorKinds) {
+        this.pauseDetectors = Objects.requireNonNull(pauseDetectorKinds, "pause detectors kinds cannot be null");
         return this;
     }
 
@@ -84,6 +86,6 @@ public class Bootstrap {
         if (strategy == null) {
             throw new IllegalStateException("Load strategy must be configured");
         }
-        return new DefaultRoadrunner(strategy, measurementProgress, outputDir, latencyRecorder);
+        return new DefaultRoadrunner(strategy, measurementProgress, outputDir, pauseDetectors);
     }
 }
