+++ roadrunner-core/src/main/java/module-info.java
@@ -18,6 +18,7 @@ module io.roadrunner.core {
     requires io.roadrunner.hdrhistogram;
     requires io.roadrunner.output.csv;
     requires io.roadrunner.api;
+    requires io.roadrunner.latency.recording;
     requires org.slf4j;
     requires jdk.jfr;
